requirements.txt
flask-restx==1.1.0
Flask-SocketIO==5.3.6
PyYAML==6.0.1
requests==2.31.0
python-socketio==5.9.0